feat(edge): support encrypted tunnel credentials
