+++ api/crypto/hash.go
